+++ poison/poison.go
@@ -30,7 +30,6 @@ import (
 	"github.com/cossacklabs/acra/acrastruct"
 	"github.com/cossacklabs/acra/crypto"
 	"github.com/cossacklabs/acra/keystore"
-	"github.com/cossacklabs/themis/gothemis/keys"
 )
 
 // Poison records length constants
@@ -55,12 +54,19 @@ func createPoisonRecordData(dataLength int) ([]byte, error) {
 }
 
 // CreatePoisonRecord generates AcraStruct encrypted with Poison Record public key
-func CreatePoisonRecord(keystore keystore.PoisonKeyStore, dataLength int) ([]byte, error) {
+func CreatePoisonRecord(keyStore keystore.PoisonKeyStorageAndGenerator, dataLength int) ([]byte, error) {
 	data, err := createPoisonRecordData(dataLength)
 	if err != nil {
 		return nil, err
 	}
-	poisonKeypair, err := keystore.GetPoisonKeyPair()
+	poisonKeypair, err := keyStore.GetPoisonKeyPair()
+	if err == keystore.ErrKeysNotFound {
+		if err = keyStore.GeneratePoisonKeyPair(); err != nil {
+			return nil, err
+		}
+		poisonKeypair, err = keyStore.GetPoisonKeyPair()
+	}
+
 	if err != nil {
 		return nil, err
 	}
@@ -74,12 +80,18 @@ func CreatePoisonRecord(keystore keystore.PoisonKeyStore, dataLength int) ([]byt
 }
 
 // CreateSymmetricPoisonRecord generates AcraBlock encrypted with Poison Record symmetric key
-func CreateSymmetricPoisonRecord(keyStore keystore.PoisonKeyStore, dataLength int) ([]byte, error) {
+func CreateSymmetricPoisonRecord(keyStore keystore.PoisonKeyStorageAndGenerator, dataLength int) ([]byte, error) {
 	data, err := createPoisonRecordData(dataLength)
 	if err != nil {
 		return nil, err
 	}
 	symmetricKey, err := keyStore.GetPoisonSymmetricKey()
+	if err == keystore.ErrKeysNotFound {
+		if err = keyStore.GeneratePoisonSymmetricKey(); err != nil {
+			return nil, err
+		}
+		symmetricKey, err = keyStore.GetPoisonSymmetricKey()
+	}
 	if err != nil {
 		return nil, err
 	}
@@ -89,10 +101,4 @@ func CreateSymmetricPoisonRecord(keyStore keystore.PoisonKeyStore, dataLength in
 	}
 
 	return crypto.SerializeEncryptedData(acraBlock, crypto.AcraBlockEnvelopeID)
-}
-
-// RecordProcessorKeyStore interface with required methods for RecordProcessor
-type RecordProcessorKeyStore interface {
-	GetPoisonPrivateKeys() ([]*keys.PrivateKey, error)
-	GetPoisonSymmetricKeys() ([][]byte, error)
 }